Cleanup accounting terminology. Cancelled invoice are no longer displayed as "invalid".
In accounting, a cancelled invoice, i.e. one for which a reverse posting exists, is still valid. To identify cancelled invoices, the terms "valid" and "cancelled" were used which might imply that a cancelled invoice is not valid. Instead a property "was cancelled" is now displayed marking the invoices with "yes" and "no".